Add per-user logging of completed, current puzzle
the server now remembers which puzzles the user has done, and which puzzle they most recently started. This way, when a user leaves and comes back, they can return to the same puzzle.

Also, in the future, it will be possible to add a user-facing profile page with information on what they've completed.